move assimilation to models
