+++ src/ff16_strategy.cpp
@@ -1,6 +1,6 @@
 #include <plant/uniroot.h>
 #include <plant/qag.h>
-#include <plant/assimilation.h>
+#include <plant/models/assimilation.h>
 #include <plant/models/ff16_strategy.h>
 #include <RcppCommon.h> // NA_REAL
